+++ src/components/autocomplete/script.ts
@@ -15,7 +15,8 @@ import StDropdownOption from '../dropdown-option/index.vue';
 import StDropdown from '../dropdown/index.vue';
 import StDropdownScript from '../dropdown/script';
 import { DropdownBindProperties } from '../dropdown/types';
-import StInput from '../input/script';
+import StInput from '../input/index.vue';
+import StInputComp from '../input/script';
 import {
   PopperPlacement,
   TriggerType,
@@ -33,6 +34,7 @@ import {
   components: {
     StDropdown,
     StDropdownOption,
+    StInput,
   },
 })
 export default class StAutocomplete extends Vue implements ValidatableComponent<AutocompleteValue> {
@@ -259,11 +261,11 @@ export default class StAutocomplete extends Vue implements ValidatableComponent<
   }
 
   focus(): void {
-    (this.$refs.input as StInput).focus();
+    (this.$refs.input as StInputComp).focus();
   }
 
   blur(): void {
-    (this.$refs.input as StInput).blur();
+    (this.$refs.input as StInputComp).blur();
   }
 
   validateValue(): AutocompleteValue {